Opt to use if conditionals instead of hash for printby_actions to correct misbehavior
For some reason, print_student_by_school was entering loop for print_student_by_batch in the previous version. Failing debugging to determine the source of error, opted instead to use if conditionals, since there are only three actions and an explicit condition for print_student_by_school was necessary due to the join action anyway.